Add pan, zoom, and minimap to GamePlayScreen instead of WorldMapScreen
