+++ composeApp/src/commonMain/kotlin/com/defenderofegril/ui/WorldMapScreen.kt
@@ -3,7 +3,6 @@ package com.defenderofegril.ui
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.gestures.detectTransformGestures
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
@@ -13,11 +12,7 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.defenderofegril.model.LevelStatus
@@ -34,12 +29,6 @@ fun WorldMapScreen(
     var showCheatDialog by remember { mutableStateOf(false) }
     var cheatCodeInput by remember { mutableStateOf("") }
     var errorMessage by remember { mutableStateOf("") }
-    
-    // State for pan and zoom
-    var scale by remember { mutableStateOf(1f) }
-    var offsetX by remember { mutableStateOf(0f) }
-    var offsetY by remember { mutableStateOf(0f) }
-    var containerSize by remember { mutableStateOf(IntSize.Zero) }
     
     Column(
         modifier = Modifier.fillMaxSize().padding(16.dp),
@@ -60,70 +49,20 @@ fun WorldMapScreen(
                 )
         )
         
-        // Zoomable and scrollable map container
-        Box(
-            modifier = Modifier
-                .weight(1f)
-                .fillMaxWidth()
-                .onSizeChanged { containerSize = it }
+        LazyVerticalGrid(
+            columns = GridCells.Fixed(2),
+            modifier = Modifier.weight(1f).fillMaxWidth(),
+            horizontalArrangement = Arrangement.spacedBy(16.dp),
+            verticalArrangement = Arrangement.spacedBy(16.dp)
         ) {
-            // Map content with pan and zoom
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .pointerInput(Unit) {
-                        detectTransformGestures { _, pan, zoom, _ ->
-                            // Apply zoom
-                            scale = (scale * zoom).coerceIn(0.5f, 3f)
-                            
-                            // Apply pan
-                            offsetX += pan.x
-                            offsetY += pan.y
-                            
-                            // Constrain pan to keep content visible
-                            val maxOffsetX = (containerSize.width * (scale - 1) / 2).coerceAtLeast(0f)
-                            val maxOffsetY = (containerSize.height * (scale - 1) / 2).coerceAtLeast(0f)
-                            
-                            offsetX = offsetX.coerceIn(-maxOffsetX, maxOffsetX)
-                            offsetY = offsetY.coerceIn(-maxOffsetY, maxOffsetY)
+            items(worldLevels) { worldLevel ->
+                LevelCard(
+                    worldLevel = worldLevel,
+                    onClick = {
+                        if (worldLevel.status != LevelStatus.LOCKED) {
+                            onLevelSelected(worldLevel.level.id)
                         }
                     }
-                    .graphicsLayer(
-                        scaleX = scale,
-                        scaleY = scale,
-                        translationX = offsetX,
-                        translationY = offsetY
-                    )
-            ) {
-                LazyVerticalGrid(
-                    columns = GridCells.Fixed(2),
-                    modifier = Modifier.fillMaxSize(),
-                    horizontalArrangement = Arrangement.spacedBy(16.dp),
-                    verticalArrangement = Arrangement.spacedBy(16.dp)
-                ) {
-                    items(worldLevels) { worldLevel ->
-                        LevelCard(
-                            worldLevel = worldLevel,
-                            onClick = {
-                                if (worldLevel.status != LevelStatus.LOCKED) {
-                                    onLevelSelected(worldLevel.level.id)
-                                }
-                            }
-                        )
-                    }
-                }
-            }
-            
-            // Minimap - shown when zoomed in
-            if (scale > 1.1f) {
-                Minimap(
-                    scale = scale,
-                    offsetX = offsetX,
-                    offsetY = offsetY,
-                    containerSize = containerSize,
-                    modifier = Modifier
-                        .align(Alignment.BottomEnd)
-                        .padding(16.dp)
                 )
             }
         }
@@ -195,69 +134,6 @@ fun WorldMapScreen(
                 }
             }
         )
-    }
-}
-
-@Composable
-fun Minimap(
-    scale: Float,
-    offsetX: Float,
-    offsetY: Float,
-    containerSize: IntSize,
-    modifier: Modifier = Modifier
-) {
-    val minimapSize = 120.dp
-    
-    Box(
-        modifier = modifier
-            .size(minimapSize)
-            .background(Color(0xCC000000))  // Semi-transparent black background
-            .border(2.dp, Color.White)
-            .padding(4.dp)
-    ) {
-        // Map outline (represents the full map)
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color(0xFF444444))
-        )
-        
-        // Viewport indicator (shows current view)
-        if (containerSize.width > 0 && containerSize.height > 0) {
-            val viewportWidthRatio = 1f / scale
-            val viewportHeightRatio = 1f / scale
-            
-            // Calculate normalized offset (-1 to 1 range)
-            val maxOffsetX = (containerSize.width * (scale - 1) / 2).coerceAtLeast(0.01f)
-            val maxOffsetY = (containerSize.height * (scale - 1) / 2).coerceAtLeast(0.01f)
-            val normalizedOffsetX = -offsetX / maxOffsetX
-            val normalizedOffsetY = -offsetY / maxOffsetY
-            
-            // Calculate viewport position in minimap
-            val viewportX = (normalizedOffsetX * (1f - viewportWidthRatio) / 2f + (1f - viewportWidthRatio) / 2f)
-            val viewportY = (normalizedOffsetY * (1f - viewportHeightRatio) / 2f + (1f - viewportHeightRatio) / 2f)
-            
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .graphicsLayer {
-                        clip = true
-                    }
-            ) {
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth(viewportWidthRatio)
-                        .fillMaxHeight(viewportHeightRatio)
-                        .align(Alignment.TopStart)
-                        .offset(
-                            x = minimapSize * viewportX,
-                            y = minimapSize * viewportY
-                        )
-                        .background(Color(0x88FFFFFF))  // Semi-transparent white for viewport
-                        .border(1.dp, Color.White)
-                )
-            }
-        }
     }
 }
